% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APackOfTheClones-package.R
\docType{package}
\name{APackOfTheClones-package}
\alias{APackOfTheClones}
\alias{APackOfTheClones-package}
\title{APackOfTheClones: Visualization of Clonal Expansion for Single Cell Immune Profiles}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualize clonal expansion via circle-packing. 'APackOfTheClones' extends 'scRepertoire' to produce a publication-ready visualization of clonal expansion at a single cell resolution, by representing expanded clones as differently sized circles. The method was originally implemented by Murray Christian and Ben Murrell in the following immunology study: Ma et al. (2021) \doi{10.1126/sciimmunol.abg6356}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qile0317.github.io/APackOfTheClones/}
  \item \url{https://github.com/Qile0317/APackOfTheClones}
  \item \url{https://joss.theoj.org/papers/10.21105/joss.06868}
  \item Report bugs at \url{https://github.com/Qile0317/APackOfTheClones/issues}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile.yang@berkeley.edu} (\href{https://orcid.org/0009-0005-0148-2499}{ORCID}) [copyright holder]

}
\keyword{internal}
