% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apotc.R
\docType{class}
\name{apotc-class}
\alias{apotc-class}
\title{The apotc (APackOfTheClones) reduction class}
\description{
A class for storing information about T cell clonal expansion in the seurat
`reductions` attribute
}
\section{Slots}{

\describe{
\item{\code{clusters}}{The original circle packing "reduction" with no modifications,
a list of clusterlists that may include NAs.}

\item{\code{centroids}}{Centroids of the clusters which default to the same centroids
of the clusterlists, tho this also means clusters has to be iterated over
everytime the plotting function is called. but it wont be slow probably}

\item{\code{clone_sizes}}{the original unscaled clone sizes. Not sure if I should
leave it named}

\item{\code{clone_scale_factor}}{scale factor to multiply `clone_sizes` by when
running the clonal expansion plotting algorithms}

\item{\code{rad_scale_factor}}{scale factor to multiply the radii in clusterlists by
after they have all been computed to increase spacing between circles. Might
also be better in the future to instead just have a number to subtract :/}

\item{\code{cluster_colors}}{character vector indicating coloration of each cluster}
}}

\keyword{internal}
