% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApotcData.R
\docType{class}
\name{ApotcData-class}
\alias{ApotcData-class}
\title{The ApotcData class}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

An S4 class for storing information about T/B cell clonal expansion to be
used by various \link{APackOfTheClones} functions. Instances of this object type
are stored by \link{RunAPOTC} in a seurat object's \verb{@misc} slot under a list named
\code{"APackOfTheClones"}. This class is not meant to be directly instantiated,
accessed, nor modified by the user.
}
\section{Slots}{

\describe{
\item{\code{clusters}}{Clustered clones which is an R list of lists of length 5,
with each list of length 5, with the first 3 elements being numeric vectors
containing the x and y coordinates and radii of each clone in the cluster.
The fourth element is a numeric vector of length 2 indicating the centroid of
the cluster, and the fifth is the estimated cluster radius.}

\item{\code{centroids}}{Centroids of the clusters which default to the same centroids
of the clusterlists, tho this also means clusters has to be iterated over
everytime the plotting function is called. but it wont be slow probably}

\item{\code{clone_sizes}}{the original unscaled clone sizes for the samples.}

\item{\code{clone_scale_factor}}{scale factor to multiply \code{clone_sizes} by when
running the clonal expansion plotting algorithms}

\item{\code{rad_scale_factor}}{scale factor to multiply the radii in clusterlists by
after they have all been computed to increase spacing between circles. Might
also be better in the future to instead just have a number to subtract :/}

\item{\code{cluster_colors}}{character vector indicating coloration of each cluster}

\item{\code{reduction_base}}{character indicating the reduction the plotting was
based off of}

\item{\code{label_coords}}{list of numeric vectors of length two indicating the (x,y)
coordinates of each label if plotted}
}}

\keyword{internal}
