% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustAPOTC.R
\name{change_clone_scale}
\alias{change_clone_scale}
\title{Modify the clone scale factor of a seurat object APackOfTheClones
reduction}
\usage{
change_clone_scale(seurat_obj, new_clone_scale, verbose = TRUE)
}
\arguments{
\item{seurat_obj}{the seurat object to be modified}

\item{new_clone_scale}{the new \code{clone_scale_factor} to change to}

\item{verbose}{boolean indicating whether to display progress information}
}
\description{
If the clone scale factor is unsatisfactory (i.e. circles are
too big/small) after creating a clonal expansion plot with \code{APOTCPlot}, the
factor can be changed with this function. This function simply repacks all
clusters with smaller circle sizes while maintaining all other factors the
same as before.
}
\examples{
library(Seurat)
suppressPackageStartupMessages(library(APackOfTheClones))
data("mini_clonotype_data","mini_seurat_obj")

# first the APOTC pipeline has to be run
pbmc <- integrate_tcr(mini_seurat_obj, mini_clonotype_data, verbose = FALSE)
pbmc <- RunAPOTC(pbmc, clone_scale_factor = 1, verbose = FALSE)

# change clone_scale_factor to 0.5
pbmc <- change_clone_scale(pbmc, 0.5, verbose = FALSE)

}
