% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_pbmc}
\alias{combined_pbmc}
\title{Multi-sampled T-cell seurat object with integrated TCR library}
\format{
A seurat object
}
\usage{
data("combined_pbmc")
}
\description{
Generated with \link{combineSeuratExpression}, more specifically, with:
\verb{r combined_pbmc@commands[["combineSeuratExpression"]]@call_string}
}
\keyword{datasets}
