% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{count_clone_sizes}
\alias{count_clone_sizes}
\title{DEFUNCT: count the number of clonotype sizes per cell cluster in a seurat
object integrated with a TCR library}
\usage{
count_clone_sizes(...)
}
\arguments{
\item{...}{arbitrary arguments}
}
\value{
error message
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\emph{\strong{ALL v0.1.x functions are deprecated}}, and the workflow has been completely revamped - now depending on the scRepertoire v2 package - which allows for the processing of multi-sampled single cell data. Please read the vignettes with \code{browseVignettes('APackOfTheClones')}or visit https://qile0317.github.io/APackOfTheClones/
}
\keyword{internal}
