% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{count_valid_barcodes}
\alias{count_valid_barcodes}
\title{Count the number of valid integrated TCR barcodes}
\usage{
count_valid_barcodes(seurat_obj)
}
\arguments{
\item{seurat_obj}{A seurat object integrated with a t cell receptor library via \code{\link{integrate_tcr}}}
}
\value{
Returns an integer indicating the number of valid barcodes that are not NA's
}
\description{
Count the number of valid integrated TCR barcodes
}
