% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clone_sizes.R
\name{countCloneSizes}
\alias{countCloneSizes}
\title{count the number of clonotype sizes per cell cluster in a seurat
object combined with a VDJ library}
\usage{
countCloneSizes(seurat_obj, clonecall = "strict", extra_filter = NULL, ...)
}
\arguments{
\item{seurat_obj}{a seurat object combined with a VDJ library with the
\code{scRepertoire} workflow and \link{combineSeuratExpression}.}

\item{clonecall}{character. The column name in the seurat object metadata to
use. See \code{scRepertoire} documentation for more information about this
parameter that is central to both packages.}

\item{extra_filter}{character. An additional string that should be formatted
\emph{exactly} like a statement one would pass into \link[dplyr:filter]{dplyr::filter} that does
\emph{additional} filtering to cells in the seurat object - on top of the other
keyword arguments - based on the metadata. This means that it will be
logically AND'ed with any keyword argument filters. This is a more flexible
alternative / addition to the filtering keyword arguments. For example, if
one wanted to filter by the length of the amino acid sequence of TCRs, one
could pass in something like \code{extra_filter = "nchar(CTaa) - 1 > 10"}. When
involving characters, ensure to enclose with single quotes.}

\item{...}{additional keyword arguments indicating the rows corresponding to
elements in the seurat object metadata that should be filtered by.
For example, seurat_clusters = c(1, 9, 10) will filter the cells to only
those in seurat clusters 1, 9, and 10. Or, if there is some column named
"sample", then sample = c("A", "B") will conduct a run only on those
corresponding cells. A useful application for this is to run the function
many times, once for each sample, and plot them all together with \link{APOTCPlot}
down the line.}
}
\value{
A list of table objects, with the table at each index corresponding
to each cluster index. Each table's names are the clonotype name indicated
by \code{clonecall} after filtering, while the values are the actual clone sizes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get clonotype frequencies from a seurat object's meta.data slot.
}
\examples{
data("combined_pbmc")

countCloneSizes(combined_pbmc)
countCloneSizes(combined_pbmc, "aa")
countCloneSizes(combined_pbmc, "nt", orig.ident = c("P17B", "P17L"))

}
