% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_seurat_obj}
\alias{mini_seurat_obj}
\title{Artificially generated Seurat object}
\format{
A Seurat object with the following slots filled
\describe{
  \item{assays}{
  \itemize{Currently only contains one assay ("RNA" - scRNA-seq expression data)
  \item{counts - Raw expression data}
  \item{data - Normalized expression data}
  \item{scale.data - Scaled expression data}
  \item{var.features - names of the current features selected as variable}
  \item{meta.features - Assay level metadata such as mean and variance}
   }}
  \item{meta.data}{Cell level metadata}
  \item{active.assay}{Current default assay}
  \item{active.ident}{Current default idents}
  \item{graphs}{Neighbor graphs computed, currently stores the SNN}
  \item{reductions}{Dimensional reductions: PCA, UMAP, and tSNE}
  \item{version}{Seurat version used to create the object}
  \item{commands}{Command history}
}
}
\usage{
data("mini_seurat_obj")
}
\description{
A generated 'SeuratObject' of a small sc-RNAseq experiment.
Has a corresponding T-cell receptor library generated from
single cell immune profiling, named `"mini_clonotype_data"`
}
\seealso{
[mini_clonotype_data()]
}
\keyword{datasets}
