% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_labels.R
\name{ModifyLabels}
\alias{ModifyLabels}
\title{Modify APackOfTheClones cluster labels}
\usage{
ModifyLabels(
  seurat_obj,
  modify_label,
  to_new_label = NULL,
  to_nudge_by = NULL,
  to_new_coord = NULL
)
}
\arguments{
\item{seurat_obj}{Object of class `Seurat` which has had `integrate_tcr` and
`RunAPOTC` functions ran on.}

\item{modify_label}{character (vector) of labels to modify (to rename
and/or move)}

\item{to_new_label}{character (vector) of new labels to convert the labels in
`modify_label` to. The order MUST be the exact same as in modify_label}

\item{`to_nudge_by`}{list of (or one single) numeric vector of length two
indicating the (x,y) amounts to *shift* the corresponding label by in the plot}

\item{`to_new_coord`}{list of (or one single) numeric vector of length two
indicating the (x,y) amounts to *move* the corresponding label to in the plot}
}
\value{
A modified version of the input `seurat_obj` where
`@reduction[['apotc']]@labels` and `@reduction[['apotc']]@label_coords` were
accordingly modified
}
\description{
Allows simpler user customization of APackOfTheClones cluster
label names and coordinates, and modifies the changes into the input seurat
object. Changes will be visible after running `APOTCPlot` on the input
seurat object. The length of the lists/vectors must all be the same, as the
indicies dictate which labels are changed.
}
\examples{
# rename "C0" and move it to x = 1, y = 2
seurat_obj <- ModifyLabels(
    seurat_obj, modify_label = "C0",
    to_new_label = "CD4+ CTL", to_new_coord = c(1, 2)
)

# rename multiple labels
seurat_obj <- ModifyLabels(
    seurat_obj, modify_label = c("C4", "C9", "C2"),
    to_new_label = c("T cm", "T cm 2", "T fh")
)

# shift multiple labels
seurat_obj <- ModifyLabels(
    seurat_obj, modify_label = c("T cm 2", "T fh", "C12"),
    to_nudge_by = list(c(1,1), c(4,3), c(0.2,-2))
)

}
