% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_screpertoire_tools.R
\name{RemoveSamplePrefix}
\alias{RemoveSamplePrefix}
\title{Remove barcode prefixes from a seurat object list}
\usage{
RemoveSamplePrefix(sc_list, prefix_end = "auto", sep = "_")
}
\arguments{
\item{sc_list}{A \code{list} of seurat objects}

\item{sep}{placeholder}
}
\value{
The modified version of the input list of seurat objects \code{sc_list}
}
\description{
placeholder
}
\examples{
data("mini_seurat_obj", "")

}
