% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAPOTC.R
\name{RunAPOTC}
\alias{RunAPOTC}
\title{Run the APackOfTheClones method on a combined Seurat object for
downstream visualization of clonal expansion}
\usage{
RunAPOTC(
  seurat_obj,
  reduction_base = "umap",
  clone_scale_factor = "auto",
  rad_scale_factor = 0.95,
  ORDER = TRUE,
  scramble = FALSE,
  try_place = FALSE,
  repulse = FALSE,
  repulsion_threshold = 1,
  repulsion_strength = 1,
  max_repulsion_iter = 20L,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object with one or more dimension reductions. Must
already have been integrated with a T cell library via
\code{integrate_tcr(seurat_obj, tcr_df)}}

\item{reduction_base}{character. The seurat reduction to base the clonal
expansion plotting on. Defaults to \code{'umap'} but can also be \code{'tsne'} or
\code{'pca'}. If `'pca'``, the cluster coordinates will be based on PC1 and PC2.
However, generally APackOfTheClones is used for displaying UMAP and
occasionally t-SNE versions to intuitively highlight clonal expansion}

\item{clone_scale_factor}{Dictates how much to scale each circle(between 0,1)
radius when converting from clonotype counts into circles that represent
individual clonotypes. The argument defaults to the character \code{"auto"}, and
if so, the most visually pleasing factor will be estimated}

\item{rad_scale_factor}{numeric between 0 and 1. This value decreases the
radius of the smallest clones by this scale factor. And the absolute value
of this decrease will be applied to all packed circles, effectively shrinking
all circles on the spot, and introduce more constant spacing in between}

\item{ORDER}{logical. Decides if the largest clone circles should be at
cluster centroids}

\item{scramble}{logical. Decides if the clone circles within each cluster
should be randomly scrambled when plotted}

\item{try_place}{If \code{TRUE}, always minimizes distance from a newly placed
circle to the origin in the circle packing algorithm}

\item{repulse}{If \code{TRUE}, will attempt to push overlapping clusters away from
each other}

\item{repulsion_threshold}{numeric. The radius that clonal circle clusters
overlap is acceptable when repulsing}

\item{repulsion_strength}{numeric. The smaller the value the less the
clusters repulse each other per iteration, and vice versa}

\item{max_repulsion_iter}{integer. The number of repulsion iterations}

\item{verbose}{logical. Decides if visual cues print to the R console of the
progress}
}
\value{
A modified version of the input seurat object with a new object in
the \verb{@reduction} attribute named \code{apotc}, which harbors data necessary for
visualizing the clonal expansion of the cells with the \link{APOTCPlot} function
}
\description{
Computes all necessary information for an APackOfTheClones
clonal expansion plot and stores it in the seurat object at
\code{seurat_obj@reduction[['apotc']]} for use later with \code{APOTCPlot}. Gets sizes
of unique clones and utilizes a circle-packing algorithm to pack circles
representing individual clones in approximately the same dimensional
reduction (\code{reduction_base}) coordinates.
}
\details{
If the user wishes to manually customize/fix the expansion plot
generated by \code{APOTCPlot}, the circular packing information can be modified
with the \code{AdjustAPOTC} function, or alternatively accessed via the syntax
\code{seurat_obj@reduction[['apotc']]} which harbours the custom S4 object of the
class \code{apotc}. Please read the function level documentation for the mentioned
functions and class, and also the web-only user vignette for a walkthrough of
the inner workings of the package.
}
\examples{
# unfinished

}
\seealso{
\link{APOTCPlot}, \link{AdjustAPOTC}
}
