% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scballpack-package.R
\docType{package}
\name{scballpack-package}
\alias{scballpack}
\alias{scballpack-package}
\title{scballpack: Visualization Of T-cell Clonal Clusters}
\description{
A circle packing package for visualizing T-cell clonal expansion by integrating scRNAseq and TCR library data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qile0317.github.io/scballpack/}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile0317@gmail.com}

}
\keyword{internal}
