% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_screpertoire_tools.R
\name{AddSamplePrefix}
\alias{AddSamplePrefix}
\title{Add scRepertoire prefixes to barcodes of a seurat object(s) pre-integration}
\usage{
AddSamplePrefix(
  sc_list,
  samples = NULL,
  ID = NULL,
  ... = NULL,
  add_to_metadata = FALSE,
  sep = "_"
)
}
\arguments{
\item{sc_list}{A `list` of seurat objects to be modified}

\item{samples}{unfinished}

\item{ID}{unfinished}

\item{...}{unfinished}

\item{add_to_metadata}{logical. If `TRUE`, adds the entire processed clonotype data
by barcode into the seurat object meta.data dataframe.}

\item{sep}{character. Decides the string to join new sample level informations.
Defaults to `"_"`.}
}
\value{
The modified version of the input list of seurat objects `sc_list`
}
\description{
placeholder.
}
\examples{
data("mini_seurat_obj", "")

}
