% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustAPOTC.R
\name{AdjustAPOTC}
\alias{AdjustAPOTC}
\title{Adjust the paramaters of the APackOfTheClones reduction in a seurat
object}
\usage{
AdjustAPOTC(
  seurat_obj,
  new_rad_scale_factor = 0L,
  adjust_repulsion = FALSE,
  repulsion_threshold = 1,
  repulsion_strength = 1,
  max_repulsion_iter = 10L,
  relocate_cluster = NULL,
  relocation_coord = NULL,
  nudge_cluster = NULL,
  nudge_vector = NULL,
  recolor_cluster = NULL,
  new_color = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The seurat object to be adjusted}

\item{new_rad_scale_factor}{}
}
\value{
The adjusted (modified) `seurat_obj`
}
\description{
unfinished
}
