% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustAPOTC.R
\name{AdjustAPOTC}
\alias{AdjustAPOTC}
\title{Adjust the paramaters of the APackOfTheClones reduction in a seurat
object}
\usage{
AdjustAPOTC(
  seurat_obj,
  new_rad_scale_factor = NULL,
  new_clone_scale_factor = NULL,
  repulse = FALSE,
  repulsion_threshold = 1,
  repulsion_strength = 1,
  max_repulsion_iter = 10L,
  relocate_cluster = NULL,
  relocation_coord = NULL,
  nudge_cluster = NULL,
  nudge_vector = NULL,
  recolor_cluster = NULL,
  new_color = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The seurat object to be adjusted. Must have an \code{apotc}
reduction}

\item{new_rad_scale_factor}{responsible for changing the rad_scale_factor of
all circles. Can be a numerical value between 0 and 1.
unfinished}
}
\value{
The adjusted (modified) \code{seurat_obj}
}
\description{
If the user is unsatisfied with the clonal expansion plot that
was generated from \code{RunAPOTC} and \code{APOTCPlot}, this function has a range of
arguments to modify the data and/or parameters of the visualization. Note
that some of the arguments may conflict with eachother.
}
