% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APackOfTheClones-package.R
\docType{package}
\name{APackOfTheClones-package}
\alias{APackOfTheClones}
\alias{APackOfTheClones-package}
\title{APackOfTheClones: Visualization of T-Cell Clonal Expansion}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualize T-cell clonal expansion via circle-packing. 'APackOfTheClones' can integrate single-cell RNA sequencing and T-cell receptor libraries from the 10X genomics Single Cell Immune Profiling and Cell Ranger pipeline, to produce a simple, publication-ready visualization of the clonal expansion. The method was originally implemented by Murray Christian and Ben Murrell in the following immunology study: Ma et al. (2021) \doi{10.1126/sciimmunol.abg6356}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qile0317.github.io/APackOfTheClones/}
  \item Report bugs at \url{https://github.com/Qile0317/APackOfTheClones/issues/}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile.yang@berkeley.edu} [copyright holder]

}
\keyword{internal}
