% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APackOfTheClones-package.R
\docType{package}
\name{APackOfTheClones-package}
\alias{APackOfTheClones}
\alias{APackOfTheClones-package}
\title{APackOfTheClones: Visualization Of T-cell Clonal Expansion}
\description{
A circle packing package for visualizing T-cell clonal expansion by integrating scRNAseq and TCR library data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qile0317.github.io/scballpack/}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile0317@gmail.com} [copyright holder]

}
\keyword{internal}
