% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apotc_plot.R
\name{APOTCPlot}
\alias{APOTCPlot}
\title{Clonal expansion plotting}
\usage{
APOTCPlot(
  seurat_obj,
  res = 360,
  linetype = "blank",
  use_default_theme = TRUE,
  show_origin = FALSE,
  retain_axis_scales = FALSE,
  add_size_legend = TRUE,
  legend_sizes = c(1, 5, 50),
  legend_position = "top_left",
  legend_buffer = 1.5,
  legend_color = "#808080",
  legend_spacing = 0.4
)
}
\description{
Clonal expansion plotting
}
