% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APOTCPlot.R
\name{APOTCPlot}
\alias{APOTCPlot}
\title{Create clonal expansion plot after RunAPOTC()}
\usage{
APOTCPlot(
  seurat_obj,
  res = 360,
  linetype = "blank",
  use_default_theme = TRUE,
  show_origin = FALSE,
  retain_axis_scales = FALSE,
  show_labels = FALSE,
  label_size = 5,
  add_size_legend = TRUE,
  legend_sizes = c(1, 5, 50),
  legend_position = "top_left",
  legend_buffer = 1.5,
  legend_color = "#808080",
  legend_spacing = "auto",
  legend_label = "Clone sizes",
  legend_text_size = 5
)
}
\description{
Given a seurat object with an 'apotc' (APackOfTheClones) object from running
RunAPOTC(), this function will read the information and return a customizable
ggplot of the clonal expansion with a circle size legend.
}
