% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_screpertoire_tools.R
\name{ChangeSamplePrefix}
\alias{ChangeSamplePrefix}
\title{Modify existing barcode prefixes from a seurat object list}
\usage{
ChangeSamplePrefix(
  sc_list,
  samples = NULL,
  ID = NULL,
  ... = NULL,
  add_to_metadata = FALSE,
  original_sep = "_",
  sep = "_"
)
}
\arguments{
\item{sc_list}{A `list` of seurat objects}

\item{sep}{placeholder}
}
\value{
The modified version of the input list of seurat objects `sc_list`
}
\description{
placeholder
}
\examples{
data("mini_seurat_obj", "")

}
