% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{clonal_expansion_plot}
\alias{clonal_expansion_plot}
\title{DEFUNCT: Visualize T cell clonal expansion with a ball-packing plot.}
\usage{
clonal_expansion_plot(...)
}
\arguments{
\item{...}{deprecated}
}
\value{
error message
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\emph{\strong{ALL v0.1.x functions are deprecated}}, and the workflow has been completely revamped - now depending on the scRepertoire v2 package - which allows for the processing of multi-sampled single cell data. Please see the new vignette, showing a superior way to achieve the same visualizations with more features.
}
\keyword{internal}
