% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{clonal_expansion_plot}
\alias{clonal_expansion_plot}
\title{DEPRECATED: Visualize T cell clonal expansion with a ball-packing plot.}
\usage{
clonal_expansion_plot(...)
}
\value{
error message
}
\description{
As of CRAN version 1.0.0, `clonal_expansion_plot` has been deprecated.
APackOfTheClones has been revamped and improved to allow for multiple
samples. Please read the new vignette with `vignette('APOTC_workflow')` and
refer to the `RunAPOTC` and `APOTCPlot` functions.
}
