% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_expansion_plot.R
\name{clonal_expansion_plot}
\alias{clonal_expansion_plot}
\title{Visualize T cell clonal expansion with a ball-packing plot.}
\usage{
clonal_expansion_plot(
  seurat_obj,
  tcr_df,
  res = 360,
  clone_scale_factor = 0.01,
  rad_scale_factor = 1,
  ORDER = TRUE,
  try_place = FALSE,
  progbar = TRUE,
  repulse = FALSE,
  repulsion_threshold = 1,
  repulsion_strength = 0.05,
  max_repulsion_iter = 100,
  use_default_theme = TRUE,
  show_origin = FALSE,
  retain_axis_scales = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object with a UMAP reduction.}

\item{tcr_df}{Dataframe of the T cell library generated by Cell Ranger}

\item{res}{The number of points on the generated path per full circle.}

\item{clone_scale_factor}{numeric. Decides how much to scale each circle. Usually should be kept at around 0.01 to somewhat maintain UMAP structure}

\item{rad_scale_factor}{numeric. indicates how much the radii of the clones should decrease to add a slight gap between all of them. Defaults to 1 but 0.85-0.95 values are recommended.}

\item{ORDER}{logical. Decides if the largest clones should be at the cluster centroids}

\item{try_place}{If `TRUE`, always minimizes distance from a newly placed circle to the origin}

\item{progbar}{logical. Decides if visual cues print to the REPL of the packing progress}

\item{repulse}{CURRENTLY BUGGED! DO NOT SET TO `TRUE`. The intented functionality is: If `TRUE`, will attempt to push overlapping clusters away from each other.}

\item{repulsion_threshold}{numeric. The radius that cluster overlap is acceptable}

\item{repulsion_strength}{numeric. The smaller the value the less the clusters repulse each other}

\item{max_repulsion_iter}{numeric. The number of repulsion iterations.}

\item{use_default_theme}{If `TRUE`, the resulting plot will have the same theme as the seurat UMAP}

\item{show_origin}{logical. If `TRUE`, only the centers of each circle will be plotted}

\item{retain_axis_scales}{If `TRUE`, approximately maintains the axis scales of the original UMAP. However, sometimes this argument breaks a few circles so it should be set to false if this happens. The fix is in-progress}
}
\value{
Returns a ggplot2 object of the ball packing plot. Can be operated on like normal ggplot objects
}
\description{
Integrates a cell ranger T cell library into a Seurat object with a UMAP
reduction. Then gets sizes of unique clones and utilizes a circle-packing
algorithm to pack circles representing individual clones in approximately
the same UMAP coordinates and clusters into a ggplot object.
}
\examples{
\dontrun{
library(Seurat)
library(readr)
library(scballpack)

# run sc-RNAseq pipeline to produce a seurat_object

seurat_object <- RunUMAP(seurat_object)
TCR_dataframe <- read.csv("file_location/all_contig_annotations.csv")

# produce and show the ball-packing plot
ball_pack_plot <- clonal_expansion_plot(seurat_object, TCR_dataframe)
ball_pack_plot
}

}
\seealso{
\code{\link{integrate_tcr}}
}
\concept{circle_packing}
