% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_pbmc}
\alias{combined_pbmc}
\title{Multi-sampled T-cell seurat object with integrated TCR library}
\format{
A seurat object
}
\usage{
data("combined_pbmc")
}
\description{
Generated with \link{combineSeuratExpression}, more specifically, with:

combineSeuratExpression(scRepertoire::combineTCR(scRepertoire::mini_contig_list, ,     samples = c("P17B", "P17L", "P18B", "P18L", "P19B", "P19L", ,         "P20B", "P20L")), scRepertoire::scRep_example)
}
\keyword{datasets}
