% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{count_clone_sizes}
\alias{count_clone_sizes}
\title{DEPRECATED: count the number of clonotype sizes per cell cluster in a seurat
object integrated with a TCR library}
\usage{
count_clone_sizes(...)
}
\value{
error message
}
\description{
As of CRAN version 1.0.0, `count_clone_sizes` has been deprecated.
APackOfTheClones has been revamped and improved to allow for multiple
samples. Please read the new vignette with `vignette('APOTC_workflow')`.
}
