% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApotcDataIndex.R
\name{getApotcDataIds}
\alias{getApotcDataIds}
\title{Get all object ids of previous RunAPOTC runs on a seurat object}
\usage{
getApotcDataIds(seurat_obj)
}
\arguments{
\item{seurat_obj}{a seurat object that has had RunAPOTC ran on it before in
order of the functions being called.}
}
\value{
a character vector of all object ids of previous RunAPOTC runs, in
the order they were ran.
}
\description{
A convenience function to get all object ids of previous RunAPOTC run IDs
}
