% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApotcDataIndex.R
\name{getLastApotcDataId}
\alias{getLastApotcDataId}
\title{Get the object id of the last RunAPOTC run on a seurat object}
\usage{
getLastApotcDataId(seurat_obj)
}
\arguments{
\item{seurat_obj}{a seurat object that has had RunAPOTC ran on it before in
order of the functions being called.}
}
\value{
a character of the object id of the last RunAPOTC call
}
\description{
A convenience function to get the last object ids of the previous RunAPOTC
run, to be used by \link{APOTCPlot} and \link{AdjustAPOTC}
}
