% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightClones.R
\name{highlightClones,gg-method}
\alias{highlightClones,gg-method}
\title{Highlight specific clones on an APackOfTheClones ggplot}
\usage{
\S4method{highlightClones}{gg}(
  sc.data,
  sequence,
  color_each = FALSE,
  default_color = "#808080"
)
}
\value{
A ggplot object with the data modified to the highlighted colors
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

TODO
}
