% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate_tcr}
\alias{integrate_tcr}
\title{Integrate a TCR library into Seurat object metadata}
\usage{
integrate_tcr(seurat_obj, tcr_file)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{tcr_file}{`data.frame` of the T cell library generated by Cell Ranger}
}
\value{
Returns a new Seurat object with new elements in the metadata
}
\description{
Modifies a copy of the current seurat object metadata by taking all the columns
of the `all_contig_annotations.csv`, and adding new elements to `seurat_obj@meta.data`
}
\details{
Columns from cells (barcodes) that had duplicates in another row are
concatenated into strings, seperated by `__` in the metadata element. Barcodes
from the TCR library that had no matches to barcodes in the `seurat_obj` will
add `NA`s for all elements of the same index.
}
\examples{
\dontrun{
library(Seurat)
library(readr)
library(scballpack)

# run sc-RNAseq pipeline to produce a seurat_object

# read in the TCR data
TCR_dataframe <- read.csv("file_location/all_contig_annotations.csv")

# integrate the TCR data into new seurat object
integrated_seurat_object <- integrate_tcr(seurat_object, TCR_dataframe)

# or simply modify the current seurat object
seurat_object <- integrate_tcr(seurat_object, TCR_dataframe)
}

}
\references{
atakanekiz (2019) Tutorial:Integrating VDJ sequencing data with Seurat. https://www.biostars.org/p/384640/
}
\concept{sc_RNAseq}
